New docs: Lit
As requested in:

* https://trello.com/c/PVnfdeaN#comment-665fffaa2798a59513617915
* https://trello.com/c/PVnfdeaN#comment-60a7c84181482c8dce7d0f55

In a period of one year, it went from 17k to 20k stars on GitHub.

This is my first time contributing to devdocs, it took me an entire day
to learn how to write a new scraper and new filters for this
documentation, including the time spent fine-tuning the result.